pyproject.toml
[tool.poetry]
name = "electronics"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
flake8 = "^7.1.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
